Fixed that quantity is a floating point number

+++ fints4k/src/commonMain/kotlin/net/codinux/banking/fints/transactions/swift/model/Holding.kt
@@ -11,7 +11,7 @@ data class Holding(
     val isin: String?,
     val wkn: String?,
     val buyingDate: LocalDate?,
-    val quantity: Int?,
+    val quantity: Double?,
     /**
      * (Durchschnittlicher) Einstandspreis/-kurs einer Einheit des Wertpapiers
      */
